[tensorflow] Add tf2xla_proto_py build rule so that we can expose tf2xla.proto in Python
Summary: For the build rule that compiles tf models into libraries, the graph config is fed in as a `.pbtxt` file containing a `tensorflow.tf2xla.Config` protobuf. For the purpose of auto-generating tests, it would be useful for us to be able to parse this, which means we need to build the python bindings for the said protobufs. (At the moment only the C++ one is built, since that's all tfcompile needs.)

Test Plan: The rule seems to DTRT in devbox, though I'm not yet testing this end to end. I've copied it from `tensorflow/compiler/xla/BUILD` which exposes python bindings for some proto.

Reviewers: ctl, leonard

Reviewed By: leonard

Subscribers: cvml-reviews

Differential Revision: https://phabricator.dropboxer.net/D424942